pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["Henning Merklinger <henning.merklinger@gmail.com>"]
description = ""
name = "iopac2calendar"
packages = [
  {include = "iopac2calendar", from = "src"}
]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
html5lib = "^1.1"
ics = "^0.7.2"
lxml = "^4.9.3"
omegaconf = "^2.3.0"
pandas = "^2.1.2"
python = ">=3.11,<3.13"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
simple-http-server = "^0.21.1"
tomli = "^2.0.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"
pre-commit = "^3.5.0"
pytest = "^7.4.3"
